chore: cleanup + split of files
