+++ EGG9000.Bot/Commands/ContextUserCommands.cs
@@ -1,5 +1,6 @@
 using Discord.Interactions;
 using Discord.WebSocket;
+using EGG9000.Bot.Interactions;
 using EGG9000.Common.Database;
 using EGG9000.Common.Services;
 using Microsoft.EntityFrameworkCore;
@@ -10,27 +11,27 @@ using System.Threading.Tasks;
 using static EGG9000.Common.Helpers.Discord.EmbedHelpers;
 
 namespace EGG9000.Bot.Commands {
-    public class ContextUserModule(IDbContextFactory<ApplicationDbContext> dbFactory, DiscordHostedService client, ILogger<ContextUserModule> logger) : Interactions.E9KModuleBase(dbFactory) {
+    public class ContextUserModule(IDbContextFactory<ApplicationDbContext> dbFactory, DiscordHostedService client, ILogger<ContextUserModule> logger) : E9KModuleBase(dbFactory) {
         private readonly DiscordHostedService _client = client;
         private readonly ILogger<ContextUserModule> _logger = logger;
 
         [UserCommand("Userstatus")]
         [DefaultMemberPermissions(Discord.GuildPermission.CreatePrivateThreads)]
-        [Interactions.StaffOnly(Interactions.StaffTier.FarmHand)]
+        [StaffOnly(StaffTier.FarmHand)]
         public async Task Userstatus(SocketGuildUser target) {
             await UserStatusCommands._userstatus(Context.Interaction, Db, _client, _logger, target, true, false);
         }
 
         [UserCommand("Contract Settings")]
         [DefaultMemberPermissions(Discord.GuildPermission.CreatePrivateThreads)]
-        [Interactions.StaffOnly(Interactions.StaffTier.FarmHand)]
+        [StaffOnly(StaffTier.FarmHand)]
         public async Task ContractSettings(SocketGuildUser target) {
             await ContractSettingsCommands.OpenContractSettings(Context.Interaction, Db, target);
         }
 
         [UserCommand("Rockets Tracker")]
         [DefaultMemberPermissions(Discord.GuildPermission.CreatePrivateThreads)]
-        [Interactions.StaffOnly(Interactions.StaffTier.FarmHand)]
+        [StaffOnly(StaffTier.FarmHand)]
         public async Task RocketsTrackerLinks(SocketGuildUser target) {
             var dbUser = await Db.DBUsers.FirstOrDefaultAsync(x => x.DiscordId == target.Id);
             if(dbUser == null) {
