+++ EGG9000.Bot/Commands/Informational/EBHistoryCommand.cs
@@ -1,6 +1,7 @@
 using Discord;
 using Discord.Interactions;
 using Discord.WebSocket;
+using EGG9000.Bot.Interactions;
 using EGG9000.Common.Database;
 using EGG9000.Common.Database.Entities;
 using EGG9000.Common.Helpers;
@@ -29,7 +30,7 @@ namespace EGG9000.Bot.Commands.Informational {
         }
     }
 
-    public partial class EBHistoryModule(IDbContextFactory<ApplicationDbContext> dbFactory) : Interactions.E9KModuleBase(dbFactory) {
+    public partial class EBHistoryModule(IDbContextFactory<ApplicationDbContext> dbFactory) : E9KModuleBase(dbFactory) {
 
         private class TextHistoryEntry(DateOnly entryDate, string ebString, string roleString, TextHistoryEntry lastEntry = null) {
             public DateOnly EntryDate { get; set; } = entryDate;
