+++ EGG9000.Bot/Commands/NewCode.cs
@@ -1,5 +1,6 @@
 using Discord.Interactions;
 using Discord.WebSocket;
+using EGG9000.Bot.Interactions;
 using EGG9000.Common.Database;
 using EGG9000.Common.Database.Entities;
 using EGG9000.Common.Helpers;
@@ -10,12 +11,12 @@ using System.Threading.Tasks;
 using static EGG9000.Common.Helpers.Discord.EmbedHelpers;
 
 namespace EGG9000.Bot.Commands {
-    public class NewCodeModule(IDbContextFactory<ApplicationDbContext> dbFactory, DiscordSocketClient client) : Interactions.E9KModuleBase(dbFactory) {
+    public class NewCodeModule(IDbContextFactory<ApplicationDbContext> dbFactory, DiscordSocketClient client) : E9KModuleBase(dbFactory) {
         private readonly DiscordSocketClient _client = client;
 
         [SlashCommand("newcoopcode", "Generate a new co-op code, a channel will be created for the co-op")]
         [DefaultMemberPermissions(Discord.GuildPermission.ManageChannels)]
-        [Interactions.StaffOnly(Interactions.StaffTier.CluckingCoordinator)]
+        [StaffOnly(StaffTier.CluckingCoordinator)]
         public async Task NewCoopCode() {
             await Context.Interaction.DeferAsync();
             var words = new Words();
@@ -34,8 +35,8 @@ namespace EGG9000.Bot.Commands {
 
         [SlashCommand("deletecoop", "Delete co-op channel from discord and database ")]
         [DefaultMemberPermissions(Discord.GuildPermission.Administrator | Discord.GuildPermission.ManageChannels | Discord.GuildPermission.ManageRoles)]
-        [Interactions.StaffOnly(Interactions.StaffTier.Admin)]
-        [Interactions.ChannelContext(Coop = true)]
+        [StaffOnly(StaffTier.Admin)]
+        [ChannelContext(Coop = true)]
         public async Task DeleteCoop() {
             await Context.Interaction.DeferAsync();
             Db.Remove(CoopChannel);
