+++ EGG9000.Common/Database/CustomFarmStats.cs
@@ -1,0 +1,12 @@
+namespace EGG9000.Common.Database {
+    public class CustomFarmStats {
+        public double CurrentShippingRate { get; set; }
+        public double EggLayingRate { get; set; }
+        public double MaxShippingRate { get; set; }
+        public double EggValue { get; set; }
+        public double Income { get; set; }
+        public double MaxRunningBonus { get; set; }
+        public double HabSpace { get; set; }
+        public int InternalHatchery { get; set; }
+    }
+}
