+++ EGG9000.Common/Database/TrophyLevel.cs
@@ -1,0 +1,9 @@
+namespace EGG9000.Common.Database {
+    public enum TrophyLevel {
+        Bronze = 1,
+        Silver = 2,
+        Gold = 3,
+        Platinum = 4,
+        Diamond = 5,
+    }
+}
